Move min / max MOS out of side panel and into toolbar
