+++ client/packages/requisitions/src/RequestRequisition/DetailView/Toolbar/Toolbar.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import React from 'react';
 import {
   AppBarContentPortal,
   InputWithLabelRow,
@@ -9,18 +9,49 @@ import {
   Alert,
   Tooltip,
   Switch,
+  Autocomplete,
+  useConfirmationModal,
 } from '@openmsupply-client/common';
 import { InternalSupplierSearchInput } from '@openmsupply-client/system';
 import { useHideOverStocked, useRequest } from '../../api';
 
-export const Toolbar: FC = () => {
+const MONTHS = [1, 2, 3, 4, 5, 6];
+
+export const Toolbar = () => {
   const { on, toggle } = useHideOverStocked();
   const t = useTranslation();
   const isDisabled = useRequest.utils.isDisabled();
   const isProgram = useRequest.utils.isProgram();
   const { itemFilter, setItemFilter } = useRequest.line.list();
-  const { theirReference, update, otherParty, programName } =
-    useRequest.document.fields(['theirReference', 'otherParty', 'programName']);
+  const {
+    minMonthsOfStock,
+    maxMonthsOfStock,
+    theirReference,
+    update,
+    otherParty,
+    programName,
+  } = useRequest.document.fields([
+    'theirReference',
+    'otherParty',
+    'programName',
+    'minMonthsOfStock',
+    'maxMonthsOfStock',
+  ]);
+
+  const getMinMOSConfirmation = useConfirmationModal({
+    title: t('heading.are-you-sure'),
+    message: t('messages.changing-min-mos'),
+  });
+
+  const getMinMOSUnassignConfirmation = useConfirmationModal({
+    title: t('heading.are-you-sure'),
+    message: t('messages.unassign-min-mos'),
+  });
+
+  const getMaxMOSConfirmation = useConfirmationModal({
+    title: t('heading.are-you-sure'),
+    message: t('messages.changing-max-mos'),
+  });
 
   return (
     <AppBarContentPortal
@@ -45,6 +76,75 @@ export const Toolbar: FC = () => {
               }
             />
           )}
+          <InputWithLabelRow
+            label={t('label.min-months-of-stock')}
+            Input={
+              <Autocomplete
+                disabled={isDisabled || isProgram}
+                clearIcon={null}
+                isOptionEqualToValue={(a, b) => a.value === b.value}
+                value={
+                  minMonthsOfStock === 0
+                    ? { label: t('label.not-set'), value: 0 }
+                    : {
+                        label: t('label.number-months', {
+                          count: minMonthsOfStock,
+                        }),
+                        value: minMonthsOfStock,
+                      }
+                }
+                width="150px"
+                options={[
+                  { label: t('label.not-set'), value: 0 },
+                  ...MONTHS.map(numberOfMonths => ({
+                    label: t('label.number-months', { count: numberOfMonths }),
+                    value: numberOfMonths,
+                  })),
+                ]}
+                onChange={(_, option) => {
+                  if (option && option.value === 0) {
+                    getMinMOSUnassignConfirmation({
+                      onConfirm: () =>
+                        update({ minMonthsOfStock: option.value }),
+                    });
+                  } else {
+                    option &&
+                      getMinMOSConfirmation({
+                        onConfirm: () =>
+                          update({ minMonthsOfStock: option.value }),
+                      });
+                  }
+                }}
+                getOptionDisabled={option => option.value > maxMonthsOfStock}
+              />
+            }
+          />
+          <InputWithLabelRow
+            label={t('label.max-months-of-stock')}
+            Input={
+              <Autocomplete
+                disabled={isDisabled || isProgram}
+                clearIcon={null}
+                isOptionEqualToValue={(a, b) => a.value === b.value}
+                value={{
+                  label: t('label.number-months', { count: maxMonthsOfStock }),
+                  value: maxMonthsOfStock,
+                }}
+                width="150px"
+                options={MONTHS.map(numberOfMonths => ({
+                  label: t('label.number-months', { count: numberOfMonths }),
+                  value: numberOfMonths,
+                }))}
+                onChange={(_, option) =>
+                  option &&
+                  getMaxMOSConfirmation({
+                    onConfirm: () => update({ maxMonthsOfStock: option.value }),
+                  })
+                }
+                getOptionDisabled={option => option.value < minMonthsOfStock}
+              />
+            }
+          />
           <InputWithLabelRow
             label={t('label.supplier-ref')}
             Input={
